Fix lint, types, and stale tests; sync docs after the P0 gate.
Keep ruff/mypy green, repair NeilAgentError and JUnit collection bugs, and align CLI/Web tests with subtask tools and parent_run_id. Leave the GitHub Actions workflow file unpushed until a token with workflow scope is available.

+++ src/neil_agent/providers/retry.py
@@ -47,7 +47,7 @@ class RetryPolicy:
             raise ValueError("retry number must be positive")
         if error.retry_after is not None:
             return min(error.retry_after, self.max_delay)
-        exponential = self.base_delay * (2 ** (retry_number - 1))
+        exponential = float(self.base_delay * (2 ** (retry_number - 1)))
         return min(exponential, self.max_delay)
